requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
ccxt==4.1.87
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
fonttools==4.47.0
frozendict==2.3.10
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.2
html5lib==1.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pandas-ta==0.3.14b0
peewee==3.17.0
Pillow==10.1.0
plotly==5.18.0
pycares==4.4.0
pycparser==2.21
pyparsing==3.1.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
yfinance==0.2.32
zope.event==5.0
zope.interface==6.1